# backend/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Academia-Industry Collaboration Portal - API",
    "description": "Backend Phase 1-6 API (Student, Industry, Academician, Institution)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/admin/knowledge-base/documents": {
      "get": {
        "tags": [
          "Knowledge Base Ingestion"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_admin_knowledge_base_documents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Knowledge Base Ingestion"
        ],
        "summary": "Upload Document",
        "description": "Module 1: Document Upload (reuses storage service conceptually, but processes content for RAG)",
        "operationId": "upload_document_admin_knowledge_base_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_admin_knowledge_base_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/admin/knowledge-base/documents/text": {
      "post": {
        "tags": [
          "Knowledge Base Ingestion"
        ],
        "summary": "Add Text Document",
        "operationId": "add_text_document_admin_knowledge_base_documents_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeDocumentCreateText"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/admin/knowledge-base/documents/{document_id}": {
      "delete": {
        "tags": [
          "Knowledge Base Ingestion"
        ],
        "summary": "Delete Document",
        "operationId": "delete_document_admin_knowledge_base_documents__document_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/knowledge-base/documents/{document_id}/embed": {
      "post": {
        "tags": [
          "Knowledge Base Ingestion"
        ],
        "summary": "Trigger Embedding",
        "description": "Manually trigger embedding for missing chunks.",
        "operationId": "trigger_embedding_admin_knowledge_base_documents__document_id__embed_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/knowledge-base/seed": {
      "post": {
        "tags": [
          "Knowledge Base Ingestion"
        ],
        "summary": "Seed Knowledge Base",
        "description": "Module 5: Seed Content",
        "operationId": "seed_knowledge_base_admin_knowledge_base_seed_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/admin/knowledge-base/test-query": {
      "get": {
        "tags": [
          "Knowledge Base Ingestion"
        ],
        "summary": "Test Query",
        "description": "Module 6: Ingestion Quality Check",
        "operationId": "test_query_admin_knowledge_base_test_query_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          },
          {
            "name": "access_level",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Access Level"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_document_admin_knowledge_base_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "document_type": {
            "type": "string",
            "title": "Document Type"
          },
          "access_level": {
            "type": "string",
            "title": "Access Level"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "default": "uploaded_file"
          }
        },
        "type": "object",
        "required": [
          "file",
          "title",
          "document_type",
          "access_level"
        ],
        "title": "Body_upload_document_admin_knowledge_base_documents_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KnowledgeDocumentCreateText": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "document_type": {
            "type": "string",
            "title": "Document Type"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "default": "manual"
          },
          "access_level": {
            "type": "string",
            "title": "Access Level"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content",
          "document_type",
          "access_level"
        ],
        "title": "KnowledgeDocumentCreateText"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}